Refactor for API structure
