+++ .env
@@ -1,6 +1,6 @@
 # This is a sample .ENV file for Beacon
 debug = false
-extern_port = "8080"
+extern_port = 3030
 cache_url = "redis://127.0.0.1"
 cache_port = "6379"
 cache_password = ""
