+++ src/api/mod.rs
@@ -1,0 +1,4 @@
+pub mod utils;
+pub mod routes;
+pub mod handlers;
+pub mod interfaces;
