+++ src/constants.rs
@@ -1,6 +1,8 @@
+/// ==== CONFIG ==== ///
+
 pub const CONFIG_FILE: &str = ".env";
 pub const SETTINGS_DEBUG: bool = false;
-pub const SETTINGS_EXTERN_PORT: &str = "3030";
+pub const SETTINGS_EXTERN_PORT: u16 = 8080;
 pub const SETTINGS_DB_URL: &str = "mongodb://127.0.0.1";
 pub const SETTINGS_DB_PORT: &str = "12701";
 pub const SETTINGS_DB_PASSWORD: &str = "password";
@@ -8,10 +10,14 @@ pub const SETTINGS_CACHE_URL: &str = "redis://127.0.0.1";
 pub const SETTINGS_CACHE_PORT: &str = "6379";
 pub const SETTINGS_CACHE_PASSWORD: &str = "password";
 
-/// ==== DRUID CONSTANTS ==== ///
+/// ==== DRUID ==== ///
 
 pub const DRUID_CHARSET: &[u8] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZ\
                             abcdefghijklmnopqrstuvwxyz\
                             0123456789";
 
 pub const DRUID_LENGTH: usize = 16;
+
+/// ==== STORAGE ==== ///
+
+pub const DB_KEY: &str = "default";
